Only call Consumer.finalize after all files have been processed
finalize can trigger vocabulary building and subword learning so it
should happen after all files have been processed.

So updating the summary of a single file should happen in the __call__
method instead. The summary is now passed when setting the context for
the active file.

This changes also introduces context manager to properly close files,
even when exceptions are raised.