Decouple consumers from the pipeline build state
Prior to this change, the pipeline should be built before using the
consumer. It seems more flexible to decouple the two objects. For
example, parallelization may require to build pipelines in separates
processes and the pipeline may not be accessible at this point.

The pipeline build_state was currently used to decide whether
alignments should be saved on disk or not. We could alternatively
directly look in the configuration.

This change also moves the update of "num_samples" in the base class
so that all calls to generate_preprocessed_data return a valid value.