Define the file context in the Consumer implementation
This change simplifies the Consumer usage by not requiring to set a
file context before passing some TUs. This is especially useful to
allow cross-file parallelization.

The batch metadata contain information about the file where the batch
is coming from, so we can use this information to know where to write
the outputs. When we get a batch from a file we did not see before,
the output files are cleared and then opened in append mode.

Consequently, the flag to write alignments on disk has been moved to
the batch metadata.