Parallelize preprocessing (#183)
When enabled, a pool of workers will apply the pipeline on multiple
batches (possibly from different files) in parallel.

The main thread is responsible of consuming the results as they are
available.

An important note is that we use multiprocessing, not
multithreading. This is required to enable true parallelism in the
Python world. The consequences are:

* The pipeline is duplicated in each process: it might be an issue if
  some operators need to load a lof of resources in their constructor.
* Logs could get mixed up: extra work way be needed not ensure that
  logs are coherent.