Add Token objects in the Tokenization structure (#198)
* Add Token objects in the Tokenization structure

The Token class makes it easier to manipulate tokens with annotations.

The replace_token method has been updated to use these objects. A side
effect is that it can no longer apply the modification in-place
because it needs to invalidate the other representation by calling the
"tok" setter. The test has been updated to not make the assumptions
that the tokens list is modified in-place.

The performance impact of this change should be small: the Token
objects are always used as an intermediate representation during
tokenization. The memory usage is of course doubled. This should not
be an issue with proper batching of the corpus.

* Only keep Token objects in TU