Add a mechanism to share operator resources across workers
Operators can declare resources that will be shared across
workers. This is needed for operators using very large
resources (e.g. alignment models).

We use the Manager concept from the multiprocessing module to build a
shared instance and pass a proxy instance to each worker. The proxy
instance can be used like the original object but the computation will
happen in a separate process.