Remove inference process methods from the Framework
In most cases, it is easier and safer to directly call a method of a
Processor instance.

The Processor.process_input is updated to rebuild the pipeline if the
example has its own configuration (V1 mode). A future PR will disable
this rebuild for V2 configuration and only accepts options.

The process_input method also needs to be thread-safe, as the
inference server is starting a new thread on each request.