Fixes vocabulary management
Different functions had different ways of parsing the vocabulary file,
which could cause inconsistencies. The parsing has been moved to
the tokenizer.vocabulary_iterator function and made more robust.

This PR also fixes the vocabulary format that is passed to the OpenNMT
Tokenizer. It does not accept header lines and float frequencies.