Always forward configuration overrides to InferenceProcessor (#293)
* Always forward configuration overrides to InferenceProcessor

Before this change, configuration overrides were not allowed when
using a V2 configuration. However, to keep compatibility with existing
inference requests we might want to accept such requests.

In this change, the serving code no longer applies the configuration
override but instead passes it to the InferenceProcessor. The
processor can then decide to:

* apply the override (e.g. a V1 configuration)
* map known overrides to operators options (e.g. V2 configuration with
  legacy overrides)
* reject the override (e.g. V2 configuration with unsupported
  overrides)

* Fix some CodeFactor issues